tests/submission/Google/audio/error/error_src_not_supported.html

<!doctype html>
<html>
 <head>
  <title>audio.error - MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-error">spec reference</a></p>
  <audio id="a" src="." onerror="test_error()">
  </audio>
  <div id="log"></div>
  <script>
function test_error() {
 var v = document.getElementById("a");
 test(function() {
  assert_equals(
   v.error.code,
   v.error.MEDIA_ERR_SRC_NOT_SUPPORTED,
   "audioElement.error property is MEDIA_ERR_SRC_NOT_SUPPORTED on bogus source");
});
}
  </script>
 </body>
</html>
